refactor: split nbd-cow benchmark modules
