+++ crates/nbd-cow/Cargo.toml
@@ -3,12 +3,20 @@ name = "nbd-cow"
 version = "0.4.42"
 edition.workspace = true
 
+[[bin]]
+name = "bench"
+path = "src/bin/bench/main.rs"
+required-features = ["bench"]
+
+[features]
+bench = ["dep:serde_json", "dep:tempfile"]
+
 [dependencies]
 bitvec = { workspace = true }
 libc = { workspace = true }
 nix = { workspace = true, features = ["fs", "socket"] }
-serde_json = { workspace = true }
-tempfile = { workspace = true }
+serde_json = { workspace = true, optional = true }
+tempfile = { workspace = true, optional = true }
 thiserror = { workspace = true }
 tokio = { workspace = true, features = ["full"] }
 tokio-util = { workspace = true }
@@ -16,6 +24,7 @@ tracing = { workspace = true }
 
 [dev-dependencies]
 nix = { workspace = true, features = ["user"] }
+tempfile = { workspace = true }
 
 [lints]
 workspace = true
