+++ crates/nbd-cow/src/bin/bench/backends/mod.rs
@@ -1,0 +1,50 @@
+mod dm_snapshot;
+mod nbd_cow;
+
+pub(crate) use dm_snapshot::run_dm_snapshot_bench;
+pub(crate) use nbd_cow::run_nbd_cow_bench;
+
+fn result_after_cleanup<T>(
+    result: Result<T, String>,
+    cleanup_errors: Vec<String>,
+) -> Result<T, String> {
+    if cleanup_errors.is_empty() {
+        return result;
+    }
+
+    let cleanup_message = cleanup_errors.join("; ");
+    match result {
+        Ok(_) => Err(format!("cleanup failed: {cleanup_message}")),
+        Err(err) => Err(format!("{err}; cleanup also failed: {cleanup_message}")),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn result_after_cleanup_returns_original_result_when_cleanup_succeeds() {
+        assert_eq!(result_after_cleanup(Ok(7), Vec::new()).unwrap(), 7);
+        assert_eq!(
+            result_after_cleanup::<u8>(Err("fio failed".to_string()), Vec::new()).unwrap_err(),
+            "fio failed"
+        );
+    }
+
+    #[test]
+    fn result_after_cleanup_reports_cleanup_errors() {
+        assert_eq!(
+            result_after_cleanup(Ok(7), vec!["remove failed".to_string()]).unwrap_err(),
+            "cleanup failed: remove failed"
+        );
+        assert_eq!(
+            result_after_cleanup::<u8>(
+                Err("fio failed".to_string()),
+                vec!["destroy failed".to_string(), "detach failed".to_string()],
+            )
+            .unwrap_err(),
+            "fio failed; cleanup also failed: destroy failed; detach failed"
+        );
+    }
+}
